+++ components/o-forms/stories/radioBtns.stories.tsx
@@ -1,5 +1,3 @@
-import withHtml from "origami-storybook-addon-html";
-import { withDesign } from "storybook-addon-designs";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import { useEffect } from "react";
 import { RadioBtn, RadioBtns } from "../src/tsx/o-forms";
@@ -26,7 +24,6 @@ const themeControl =
 export default {
 	title: "Components/o-forms/radio-buttons",
 	component: RadioBtns,
-	decorators: [withDesign, withHtml],
 	argTypes: {
 		children: hideArg,
 		theme: themeControl,