+++ components/o-forms/stories/select.stories.tsx
@@ -1,5 +1,3 @@
-import withHtml from 'origami-storybook-addon-html';
-import {withDesign} from 'storybook-addon-designs';
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 import {useEffect} from 'react';
 import './forms.scss';
@@ -16,7 +14,6 @@ const hideArg = {
 export default {
 	title: 'Components/o-forms/select-box',
 	component: Select,
-	decorators: [withDesign, withHtml],
 	argTypes: {
 		onChange: hideArg,
 		suffix: hideArg,