+++ components/o-forms/stories/toggleCheckboxes.stories.tsx
@@ -1,5 +1,3 @@
-import withHtml from 'origami-storybook-addon-html';
-import {withDesign} from 'storybook-addon-designs';
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 import {useEffect} from 'react';
 import {Checkbox, Checkboxes} from '../src/tsx/o-forms';
@@ -14,7 +12,6 @@ const hideArg = {
 export default {
 	title: 'Components/o-forms/toggle-checkboxes',
 	component: Checkboxes,
-	decorators: [withDesign, withHtml],
 	argTypes: {
 		children: hideArg,
 		theme: hideArg