+++ components/o-icons/stories/icons.stories.tsx
@@ -1,6 +1,4 @@
 import type {ComponentMeta, ComponentStory} from '@storybook/react';
-import withHtml from 'origami-storybook-addon-html';
-import {withDesign} from 'storybook-addon-designs';
 import {icons} from '../src/tsx/data';
 import {Icon} from '../src/tsx/icon';
 import './icons.scss';
@@ -8,7 +6,6 @@ import './icons.scss';
 export default {
 	title: 'Components/o-icons',
 	component: Icon,
-	decorators: [withDesign, withHtml],
 	argTypes: {
 		icon: {
 			control: 'select',