+++ components/o-labels/stories/internal/service-tier.stories.tsx
@@ -1,5 +1,3 @@
-import withHtml from 'origami-storybook-addon-html';
-import {withDesign} from 'storybook-addon-designs';
 import {ServiceLabel} from '../../src/tsx/label';
 import '../labels.scss';
 
@@ -19,8 +17,7 @@ const ComponentDescription = {
 			  type: 'text'
 			}
 		  }
-	},
-	decorators: [withDesign, withHtml]
+	}
 };
 
 export default ComponentDescription;
@@ -31,4 +28,7 @@ export const ServiceTierLabel = args => {
 		delete args.size;
 	}
 	return <ServiceLabel {...args}>{copy}</ServiceLabel>;
+}
+ServiceTierLabel.args = {
+	text: 'Bronze'
 }